Removed artifacts from an outdated makenek
The SIZE files for eddy_neknek and mv_cyl contained some variable
definitions that appeared to be added by an older version of makenek.
For example, a definition of cv_maxl was in eddy_neknek/SIZE with the
comment "automatically added by makenek".  The current version of
maknek does NOT add that variable.  Furthermore, it conflicts with the
variable definition in core/CVODE.